# Eliminates Dart->Host response copy for large buffers