--- a/lib/ui/BUILD.gn
+++ b/lib/ui/BUILD.gn
@@ -231,6 +231,7 @@ if (enable_unittests) {
       "painting/vertices_unittests.cc",
       "semantics/semantics_update_builder_unittests.cc",
       "window/platform_configuration_unittests.cc",
+      "window/platform_message_response_dart_unittests.cc",
       "window/pointer_data_packet_converter_unittests.cc",
     ]