## description
This constitutes a 30% reduction in the [1MB test](https://github.com/flutter/flutter/blob/master/dev/benchmarks/platform_channels_benchmarks/lib/main.dart) and little change in the other tests.

This will affect platform channel communications, including loading assets from the asset directory.

This is an alternative to the PR that eliminates copies if we have a modifiable buffer: https://github.com/flutter/engine/pull/34018

## breaking change notice

This can break users at runtime in the following cases:
 1) usage of a BinaryCodec that edits the ByteData returned from platform message calls
 2) custom implementations of MessageCodec that edit the ByteData returned from platform message calls

It isn't technically a breaking change since it doesn't appear to break any existing tests.

## local testing results
Local iOS platform channel benchmark results with the framework pr https://github.com/flutter/flutter/pull/109707:
```
## before ##
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Small: 47.8 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Large: 647.6 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/Large: 52.8 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/1MB: 485.9 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/SmallParallel3: 25.5 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/Small: 39.2 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/SmallParallel3: 22.5 µs

## after ##
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Small: 43.0 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Large: 660.3 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/Large: 47.9 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/1MB: 342.4 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/SmallParallel3: 26.4 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/Small: 38.6 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/SmallParallel3: 21.4 µs
```


## Pre-launch Checklist

- [x] I read the [Contributor Guide] and followed the process outlined there for submitting PRs.
- [x] I read the [Tree Hygiene] wiki page, which explains my responsibilities.
- [x] I read and followed the [Flutter Style Guide] and the [C++, Objective-C, Java style guides].
- [ ] I listed at least one issue that this PR fixes in the description above.
- [ ] I added new tests to check the change I am making or feature I am adding, or Hixie said the PR is test-exempt. See [testing the engine] for instructions on
writing and running engine tests.
- [x] I updated/added relevant documentation (doc comments with `///`).
- [x] I signed the [CLA].
- [x] All existing and new tests are passing.

If you need help, consider asking for advice on the #hackers-new channel on [Discord].

<!-- Links -->
[Contributor Guide]: https://github.com/flutter/flutter/wiki/Tree-hygiene#overview
[Tree Hygiene]: https://github.com/flutter/flutter/wiki/Tree-hygiene
[Flutter Style Guide]: https://github.com/flutter/flutter/wiki/Style-guide-for-Flutter-repo
[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style
[testing the engine]: https://github.com/flutter/flutter/wiki/Testing-the-engine
[CLA]: https://cla.developers.google.com/
[flutter/tests]: https://github.com/flutter/tests
[breaking change policy]: https://github.com/flutter/flutter/wiki/Tree-hygiene#handling-breaking-changes
[Discord]: https://github.com/flutter/flutter/wiki/Chat