@gaaclarke 2022-08-17
@a-siva @rmacnak-google I tested the performance of the UnmodifiableByteDataView, looks good to me.  See the results in the description.  Thanks!  I'll see if I can get the UnmodifiableByteDataView constructor called from c++ to avoid the framework change and put this up for review.