@gaaclarke 2022-08-17
I've tried to eliminate the change in the framework by allocating the UnmodifiableByteDataView in the c++ code but Dart_New crashes with "unreachable".

```c++
void PlatformMessageResponseDart::Complete(std::unique_ptr<fml::Mapping> data) {
  PostCompletion(
      std::move(callback_), ui_task_runner_, &is_complete_, channel_,
      [data = std::move(data)]() mutable {
        Dart_Handle byte_buffer;
        intptr_t size = data->GetSize();
        if (data->GetSize() > 1000) {
          // The ByteData will be wrapped with an
          // UnmodifiableByteDataView to protect constness in Dart.
          void* mapping = const_cast<uint8_t*>(data->GetMapping());
          byte_buffer = Dart_NewExternalTypedDataWithFinalizer(
              /*type=*/Dart_TypedData_kByteData,
              /*data=*/mapping,
              /*length=*/size,
              /*peer=*/data.release(),
              /*external_allocation_size=*/size,
              /*callback=*/MappingFinalizer);
        } else {
          byte_buffer =
              tonic::DartByteData::Create(data->GetMapping(), data->GetSize());
        }

        Dart_Handle type_data = Dart_LookupLibrary(
            tonic::DartConverter<std::string>().ToDart("dart:typed_data"));
        FML_DCHECK(!(Dart_IsNull(type_data) || Dart_IsError(type_data)));
        Dart_Handle unmodifiable_byte_data_type =
            Dart_GetNullableType(type_data,
                                 tonic::DartConverter<std::string>().ToDart(
                                     "UnmodifiableByteDataView"),
                                 0, nullptr);
        FML_DCHECK(!(Dart_IsNull(unmodifiable_byte_data_type) ||
                     Dart_IsError(unmodifiable_byte_data_type)));
        return Dart_New(unmodifiable_byte_data_type, Dart_Null(), 1,
                        &byte_buffer);
      });
}
```