@gaaclarke 2022-08-17
I've also attempted adding the unmodifiedbytedataview allocation in the PlatformDispatcher, but it incurs a higher cost.

```
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Small: 51.0 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Large: 665.4 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/Large: 49.8 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/1MB: 348.0 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/SmallParallel3: 24.9 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/Small: 40.0 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/SmallParallel3: 22.7 µs
```

```diff
--- a/lib/ui/platform_dispatcher.dart
+++ b/lib/ui/platform_dispatcher.dart
@@ -521,14 +521,17 @@ class PlatformDispatcher {
   /// The framework invokes [callback] in the same zone in which this method was
   /// called.
   void sendPlatformMessage(String name, ByteData? data, PlatformMessageResponseCallback? callback) {
+    PlatformMessageResponseCallback? mutableCallback = callback == null ? null : (ByteData? mutableData) {
+      callback(mutableData == null ? null : UnmodifiableByteDataView(mutableData));
+    };
     final String? error =
-        _sendPlatformMessage(name, _zonedPlatformMessageResponseCallback(callback), data);
+        _sendPlatformMessage(name, _zonedPlatformMessageResponseCallback(mutableCallback), data);
     if (error != null) {
       throw Exception(error);
     }
   }
```