@gaaclarke 2022-08-17
I tried wrapping the constructor for UnmodifiableByteDataView into a function in dart:ui, this also provides tiny overhead for the standard case as well.

```
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Small: 53.4 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Large: 667.3 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/Large: 50.7 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/1MB: 347.7 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/SmallParallel3: 24.9 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/Small: 44.4 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/SmallParallel3: 22.7 µs
```

```c++
void PlatformMessageResponseDart::Complete(std::unique_ptr<fml::Mapping> data) {
  PostCompletion(std::move(callback_), ui_task_runner_, &is_complete_, channel_,
                 [data = std::move(data)]() mutable {
                   Dart_Handle byte_buffer;
                   intptr_t size = data->GetSize();
                   if (data->GetSize() > 1000) {
                     // The ByteData will be wrapped with an
                     // UnmodifiableByteDataView to protect constness in Dart,
                     // see //lib/ui/platform_dispatcher.dart
                     void* mapping = const_cast<uint8_t*>(data->GetMapping());
                     byte_buffer = Dart_NewExternalTypedDataWithFinalizer(
                         /*type=*/Dart_TypedData_kByteData,
                         /*data=*/mapping,
                         /*length=*/size,
                         /*peer=*/data.release(),
                         /*external_allocation_size=*/size,
                         /*callback=*/MappingFinalizer);
                   } else {
                     byte_buffer = tonic::DartByteData::Create(
                         data->GetMapping(), data->GetSize());
                   }
                   Dart_Handle ui_lib = Dart_LookupLibrary(
                       tonic::DartConverter<std::string>().ToDart("dart:ui"));
                   FML_DCHECK(!(Dart_IsNull(ui_lib) || Dart_IsError(ui_lib)));
                   Dart_Handle result =
                       Dart_Invoke(ui_lib,
                                   tonic::DartConverter<std::string>().ToDart(
                                       "_wrapUnmodifiableByteData"),
                                   1, &byte_buffer);
                   FML_DCHECK(!(Dart_IsNull(result) || Dart_IsError(result)));
                   return result;
                 });
```