@gaaclarke 2022-08-17
Here are the results of selectively wrapping only when we are are using external typed data:
```
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Small: 50.5 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Large: 670.4 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/Large: 52.0 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/1MB: 340.2 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/SmallParallel3: 24.6 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/Small: 39.2 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/SmallParallel3: 22.2 µs
```