@gaaclarke 2022-08-17
The results between always making the UnmodifiableByteDataView and selectively doing it are so so similar I'm going to go with always creating it to make the behavior more reliable.