@rmacnak-google 2022-08-17
Note the existence of `_wrapUnmodifiableByteData` as an entry point means the unmodifiable views cannot be tree-shaken in all Flutter programs. This in turn means the AOT compiler will not be able to eliminate the is-mutable checks before typed data writes if unmodifiable views are never used (an optimization that does not currently exist). If you find the level of performance without this optimization acceptable, the VM could offer a `Dart_NewUnmodifiableTypedData` variant of `Dart_NewExternalTypedDataWithFinalizer`, which would prevent shaking the unmodifiable views for all Dart programs and forgo the possibility of this optimization. The VM would be able to label unmodifiable views created in this was as pass-by-reference within an isolate group.

@mkustermann